rename useTrainModel to useTrainModelState
