+++ application/ui/src/features/models/train-model/select-dataset-revision.component.tsx
@@ -3,10 +3,10 @@
 
 import { Content, ContextualHelp, Heading, Item, Picker } from '@geti/ui';
 
-import { useTrainModel } from './train-model-provider.component';
+import { useTrainModelState } from './train-model-provider.component';
 
 export const SelectDatasetRevision = () => {
-    const { datasetRevisions, selectedDatasetRevisionId, onSelectDatasetRevisionId } = useTrainModel();
+    const { datasetRevisions, selectedDatasetRevisionId, onSelectDatasetRevisionId } = useTrainModelState();
 
     return (
         <>
@@ -20,7 +20,7 @@ export const SelectDatasetRevision = () => {
                     <ContextualHelp variant={'info'} placement={'top'}>
                         <Heading>Selecting a dataset</Heading>
                         <Content>
-                            {`Choose the version of the dataset to use for training. If you want to train the new model 
+                            {`Choose the version of the dataset to use for training. If you want to train the new model
                             on the exact same data (media and annotations) as another model, please select the
                             corresponding dataset revision. Conversely, if you want to train on the most recent version
                             of the data (what you see in the "Dataset" page), please select "Use
